Refactors core objects for modular design
